Apply some styling to the calculator
Style html so that everything is centred and has the right font.
Flex the main calc parent element so that the calculator itself is
centred on the page with a fixed height and background colour.
Using grid, arrange the calculator buttons as desired with 4 columns
and 5 rows.
Fix the display so that it takes up its entire row. Using flexbox,
right-align the displayed number.